feat: No longer expose a property that stores the CurrentHandle inside the various script processors
Since the ForEachEntity is a const function it cannot call any latent actions, however it can still call functions that execute delegate. If those delegates are every run and use the properties, they might operate on the "wrong" one. For this reason, the entity should be sent as an optional payload which renders the "_CurrentHandle" variable useless